FISH-468 Rework admin audit upgrade.
The admin audit service isn't present in Payara Micro, leading to CNFEs
when you try to start it. This reworks the upgrade of this specific
service into its own module and service, rather than being done as a
part of each notifier upgrade. This allows us to split it out and
include it separately.

Signed-off-by: Andrew Pielage <pandrex247@hotmail.com>